+++ tests/coord_transformations/pseudo_cartesian_jacobian_matrix.cpp
@@ -6,8 +6,8 @@
 #include "circular_to_cartesian.hpp"
 #include "combined_mapping.hpp"
 #include "czarny_to_cartesian.hpp"
-#include "discrete_mapping_builder.hpp"
-#include "discrete_to_cartesian.hpp"
+#include "discrete_poloidal_cs_spline_mapping.hpp"
+#include "discrete_poloidal_cs_spline_mapping_builder.hpp"
 #include "geometry_pseudo_cartesian.hpp"
 #include "math_tools.hpp"
 
@@ -223,13 +223,17 @@ public:
                 CartesianToCircular<X_pC, Y_pC, R, Theta>>;
         const PseudoCartToCircToCart
                 pseudo_cart_to_circ_to_cart(circ_to_cart, pseudo_cart_to_circ, 1e-12);
-        DiscreteToCartesianBuilder<X, Y, SplineRThetaBuilder_host, SplineRThetaEvaluator>
+        DiscretePoloidalCSSplineMappingBuilder<
+                X,
+                Y,
+                SplineRThetaBuilder_host,
+                SplineRThetaEvaluator>
                 mapping_builder_circ(
                         Kokkos::DefaultHostExecutionSpace(),
                         circ_to_cart,
                         builder,
                         spline_evaluator);
-        DiscreteToCartesian discrete_mapping_circ_to_cart = mapping_builder_circ();
+        DiscretePoloidalCSSplineMapping discrete_mapping_circ_to_cart = mapping_builder_circ();
         using DiscreteMappingCirc = CombinedMapping<
                 decltype(discrete_mapping_circ_to_cart),
                 CartesianToCircular<X_pC, Y_pC, R, Theta>>;
@@ -258,13 +262,17 @@ public:
                 CartesianToCircular<X_pC, Y_pC, R, Theta>>;
         const PseudoCartToCzarnyToCart
                 pseudo_cart_to_czarny_to_cart(czarny_to_cart, pseudo_cart_to_circ, 1e-12);
-        DiscreteToCartesianBuilder<X, Y, SplineRThetaBuilder_host, SplineRThetaEvaluator>
+        DiscretePoloidalCSSplineMappingBuilder<
+                X,
+                Y,
+                SplineRThetaBuilder_host,
+                SplineRThetaEvaluator>
                 mapping_builder_czarny(
                         Kokkos::DefaultHostExecutionSpace(),
                         czarny_to_cart,
                         builder,
                         spline_evaluator);
-        DiscreteToCartesian discrete_mapping_czarny_to_cart = mapping_builder_czarny();
+        DiscretePoloidalCSSplineMapping discrete_mapping_czarny_to_cart = mapping_builder_czarny();
         using DiscreteMappingCzarny = CombinedMapping<
                 decltype(discrete_mapping_czarny_to_cart),
                 CartesianToCircular<X_pC, Y_pC, R, Theta>>;