+++ tests/interpolation/polar_bsplines.cpp
@@ -7,8 +7,8 @@
 #include <gtest/gtest.h>
 
 #include "circular_to_cartesian.hpp"
-#include "discrete_mapping_builder.hpp"
-#include "discrete_to_cartesian.hpp"
+#include "discrete_poloidal_cs_spline_mapping.hpp"
+#include "discrete_poloidal_cs_spline_mapping_builder.hpp"
 #include "polar_bsplines.hpp"
 #include "test_utils.hpp"
 #include "view.hpp"
@@ -197,13 +197,13 @@ TYPED_TEST(PolarBsplineFixture, PartitionOfUnity)
             theta_extrapolation_rule);
 
     const CircToCart coord_changer;
-    DiscreteToCartesianBuilder<X, Y, SplineRThetaBuilder_host, SplineRThetaEvaluator>
+    DiscretePoloidalCSSplineMappingBuilder<X, Y, SplineRThetaBuilder_host, SplineRThetaEvaluator>
             mapping_builder(
                     Kokkos::DefaultHostExecutionSpace(),
                     coord_changer,
                     builder_rp,
                     evaluator_rp);
-    DiscreteToCartesian mapping = mapping_builder();
+    DiscretePoloidalCSSplineMapping mapping = mapping_builder();
     ddc::init_discrete_space<BSplines>(mapping);
 
     int const n_eval = (BSplinesR::degree() + 1) * (BSplinesTheta::degree() + 1);