+++ tests/interpolation/polar_splines.cpp
@@ -8,8 +8,8 @@
 
 #include "circular_to_cartesian.hpp"
 #include "czarny_to_cartesian.hpp"
-#include "discrete_mapping_builder.hpp"
-#include "discrete_to_cartesian.hpp"
+#include "discrete_poloidal_cs_spline_mapping.hpp"
+#include "discrete_poloidal_cs_spline_mapping_builder.hpp"
 #include "polar_bsplines.hpp"
 #include "polar_spline_evaluator.hpp"
 #include "view.hpp"
@@ -171,12 +171,12 @@ TEST(PolarSplineTest, ConstantEval)
 #elif defined(CZARNY_MAPPING)
     CircToCart const coord_changer(0.3, 1.4);
 #endif
-    DiscreteToCartesianBuilder<X, Y, BuilderRTheta, EvaluatorRTheta> mapping_builder(
+    DiscretePoloidalCSSplineMappingBuilder<X, Y, BuilderRTheta, EvaluatorRTheta> mapping_builder(
             Kokkos::DefaultHostExecutionSpace(),
             coord_changer,
             builder_rtheta,
             evaluator_rtheta);
-    DiscreteToCartesian mapping = mapping_builder();
+    DiscretePoloidalCSSplineMapping mapping = mapping_builder();
     ddc::init_discrete_space<BSplines>(mapping);
 
     SplineMem coef(ddc::discrete_space<BSplines>().full_domain());
@@ -292,12 +292,12 @@ void test_polar_spline_eval_gpu()
 #elif defined(CZARNY_MAPPING)
     CircToCart const coord_changer(0.3, 1.4);
 #endif
-    DiscreteToCartesianBuilder<X, Y, BuilderRTheta, EvaluatorRTheta> mapping_builder(
+    DiscretePoloidalCSSplineMappingBuilder<X, Y, BuilderRTheta, EvaluatorRTheta> mapping_builder(
             Kokkos::DefaultExecutionSpace(),
             coord_changer,
             builder_rtheta,
             evaluator_rtheta);
-    DiscreteToCartesian mapping = mapping_builder();
+    DiscretePoloidalCSSplineMapping mapping = mapping_builder();
     ddc::init_discrete_space<BSplines>(mapping);
 
     SplineMem coef(ddc::discrete_space<BSplines>().full_domain());
@@ -411,12 +411,12 @@ void test_polar_integrals()
 #elif defined(CZARNY_MAPPING)
     CircToCart const coord_changer(0.3, 1.4);
 #endif
-    DiscreteToCartesianBuilder<X, Y, BuilderRTheta, EvaluatorRTheta> mapping_builder(
+    DiscretePoloidalCSSplineMappingBuilder<X, Y, BuilderRTheta, EvaluatorRTheta> mapping_builder(
             Kokkos::DefaultExecutionSpace(),
             coord_changer,
             builder_rtheta,
             evaluator_rtheta);
-    DiscreteToCartesian mapping = mapping_builder();
+    DiscretePoloidalCSSplineMapping mapping = mapping_builder();
     ddc::init_discrete_space<BSplines>(mapping);
 
     SplineMem bspline_integrals_alloc(ddc::discrete_space<BSplines>().full_domain());